pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ffdecmcp"
version = "0.1.0"
description = "MCP wrapper for JPEXS Free Flash Decompiler (FFDec) - expose SWF decompilation tools to AI assistants"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "sublimnl" }
]
keywords = ["mcp", "ffdec", "swf", "flash", "decompiler", "actionscript"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastmcp>=2.14.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/sublimnl/ffdecmcp"
Repository = "https://github.com/sublimnl/ffdecmcp"
Issues = "https://github.com/sublimnl/ffdecmcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
ffdecmcp = "ffdecmcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ffdecmcp = ["py.typed"]
